general refactor
